# contrato/openapi.json
{
  "components": {
    "schemas": {
...
      "BucketConfigurado": {
        "properties": {
          "accion": {
            "title": "Accion",
            "type": "string"
          },
          "activo": {
            "title": "Activo",
            "type": "boolean"
          },
          "alerta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerta"
          },
          "codigo": {
            "title": "Codigo",
            "type": "string"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "desde": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desde"
          },
          "etiqueta": {
            "title": "Etiqueta",
            "type": "string"
          },
          "hasta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hasta"
          },
          "orden": {
            "title": "Orden",
            "type": "integer"
          },
          "prioridad_base": {
            "title": "Prioridad Base",
            "type": "integer"
          },
          "prioridad_etiqueta": {
            "title": "Prioridad Etiqueta",
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "etiqueta",
          "desde",
          "hasta",
          "color",
          "prioridad_base",
          "prioridad_etiqueta",
          "accion",
          "alerta",
          "orden",
          "activo"
        ],
        "title": "BucketConfigurado",
        "type": "object"
      },
      "CambioParametro": {
        "description": "Peticion para fijar un umbral.\n\nEs el camino por el que R01 y R02 se activan: sin despliegue, con rastro.\nEl usuario no viaja aqui: sale del token, porque un rastro que el cliente\nfirma a su gusto no es un rastro.",
        "properties": {
          "valor": {
            "description": "Valor nuevo. Se guarda como texto y la regla lo interpreta.",
            "title": "Valor",
            "type": "string"
          }
        },
        "required": [
          "valor"
        ],
        "title": "CambioParametro",
        "type": "object"
      },
...
      "Configuracion": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/BucketConfigurado"
            },
            "title": "Buckets",
            "type": "array"
          },
          "empresa_id": {
            "title": "Empresa Id",
            "type": "string"
          },
          "reglas": {
            "items": {
              "$ref": "#/components/schemas/ReglaConfigurada"
            },
            "title": "Reglas",
            "type": "array"
          },
          "version_parametros": {
            "title": "Version Parametros",
            "type": "string"
          }
        },
        "required": [
          "empresa_id",
          "version_parametros",
          "buckets",
          "reglas"
        ],
        "title": "Configuracion",
        "type": "object"
      },
...
      "Ejecucion": {
        "properties": {
          "alertas_cerradas": {
            "title": "Alertas Cerradas",
            "type": "integer"
          },
          "alertas_generadas": {
            "title": "Alertas Generadas",
            "type": "integer"
          },
          "corte": {
            "format": "date",
            "title": "Corte",
            "type": "string"
          },
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "filas_procesadas": {
            "title": "Filas Procesadas",
            "type": "integer"
          },
          "fin": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fin"
          },
          "inicio": {
            "format": "date-time",
            "title": "Inicio",
            "type": "string"
          },
          "mensaje": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mensaje"
          }
        },
        "required": [
          "corte",
          "inicio",
          "fin",
          "filas_procesadas",
          "alertas_generadas",
          "alertas_cerradas",
          "estado",
          "mensaje"
        ],
        "title": "Ejecucion",
        "type": "object"
      },
      "EntradaAuditoria": {
        "properties": {
          "campo": {
            "title": "Campo",
            "type": "string"
          },
          "entidad": {
            "title": "Entidad",
            "type": "string"
          },
          "fecha_hora": {
            "format": "date-time",
            "title": "Fecha Hora",
            "type": "string"
          },
          "usuario_id": {
            "title": "Usuario Id",
            "type": "string"
          },
          "valor_anterior": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Anterior"
          },
          "valor_nuevo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Nuevo"
          }
        },
        "required": [
          "fecha_hora",
          "usuario_id",
          "entidad",
          "campo",
          "valor_anterior",
          "valor_nuevo"
        ],
        "title": "EntradaAuditoria",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PeticionEjecucion": {
        "properties": {
          "corte": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Corte a procesar. Si se omite, se usa la fecha de hoy en America/Bogota, que es la zona de referencia del motor (C-11).",
            "title": "Corte"
          }
        },
        "title": "PeticionEjecucion",
        "type": "object"
      },
      "ReglaConfigurada": {
        "properties": {
          "accion": {
            "title": "Accion",
            "type": "string"
          },
          "activa": {
            "title": "Activa",
            "type": "boolean"
          },
          "ambito": {
            "title": "Ambito",
            "type": "string"
          },
          "codigo": {
            "title": "Codigo",
            "type": "string"
          },
          "etiqueta": {
            "title": "Etiqueta",
            "type": "string"
          },
          "faltantes": {
            "items": {
              "type": "string"
            },
            "title": "Faltantes",
            "type": "array"
          },
          "motivo_inactiva": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Inactiva"
          },
          "parametros": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Parametros",
            "type": "object"
          },
          "parametros_requeridos": {
            "items": {
              "type": "string"
            },
            "title": "Parametros Requeridos",
            "type": "array"
          },
          "prioridad": {
            "title": "Prioridad",
            "type": "integer"
          }
        },
        "required": [
          "codigo",
          "etiqueta",
          "ambito",
          "parametros_requeridos",
          "parametros",
          "faltantes",
          "activa",
          "motivo_inactiva",
          "prioridad",
          "accion"
        ],
        "title": "ReglaConfigurada",
        "type": "object"
      },
      "ResultadoEjecucion": {
        "properties": {
          "alertas_actualizadas": {
            "title": "Alertas Actualizadas",
            "type": "integer"
          },
          "alertas_cerradas": {
            "title": "Alertas Cerradas",
            "type": "integer"
          },
          "alertas_insertadas": {
            "title": "Alertas Insertadas",
            "type": "integer"
          },
          "clientes": {
            "title": "Clientes",
            "type": "integer"
          },
          "corte": {
            "format": "date",
            "title": "Corte",
            "type": "string"
          },
          "empresa_id": {
            "title": "Empresa Id",
            "type": "string"
          },
          "filas_leidas": {
            "title": "Filas Leidas",
            "type": "integer"
          },
          "reglas_inactivas": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Reglas Inactivas",
            "type": "object"
          },
          "version_parametros": {
            "title": "Version Parametros",
            "type": "string"
          }
        },
        "required": [
          "empresa_id",
          "corte",
          "filas_leidas",
          "alertas_insertadas",
          "alertas_actualizadas",
          "alertas_cerradas",
          "clientes",
          "version_parametros",
          "reglas_inactivas"
        ],
        "title": "ResultadoEjecucion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nMotor de alertas de cartera de BUSINT.\n\nConvierte el aging de cuentas por cobrar en una lista de trabajo priorizada.\nToda alerta lleva la explicacion de por que se disparo: la regla, el parametro\nvigente y el valor que la disparo.\n\n**Los montos viajan como cadena**, no como numero, para conservar los dos\ndecimales sin el error de redondeo de los flotantes de JavaScript.\n\nAutenticacion: `POST /api/v1/sesion` devuelve un token; el resto de endpoints\nlo esperan en `Authorization: Bearer`. Los permisos son acumulativos segun el\nrol (Consulta, Gestor, Coordinador, Administrador).\n",
    "title": "BUSINT - Motor de alertas de cartera",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/configuracion/auditoria": {
      "get": {
        "operationId": "historial_api_v1_configuracion_auditoria_get",
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntradaAuditoria"
                  },
                  "title": "Response Historial Api V1 Configuracion Auditoria Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Historial de cambios de configuracion",
        "tags": [
          "configuracion"
        ]
      }
    },
    "/api/v1/configuracion/reglas/{codigo}/parametros/{nombre}": {
      "put": {
        "description": "§8.4: solo usuarios autorizados modifican reglas. C-13 fija cual rol.",
        "operationId": "fijar_api_v1_configuracion_reglas__codigo__parametros__nombre__put",
        "parameters": [
          {
            "in": "path",
            "name": "codigo",
            "required": true,
            "schema": {
              "title": "Codigo",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "nombre",
            "required": true,
            "schema": {
              "title": "Nombre",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioParametro"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fijar el valor de un parametro",
        "tags": [
          "configuracion"
        ]
      }
    },
...
    "/api/v1/ejecucion": {
      "get": {
        "operationId": "bitacora_api_v1_ejecucion_get",
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Ejecucion"
                  },
                  "title": "Response Bitacora Api V1 Ejecucion Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bitacora de corridas",
        "tags": [
          "ejecucion"
        ]
      },
      "post": {
        "description": "§10.2: reproceso manual. C-13 lo reserva al coordinador.",
        "operationId": "ejecutar_api_v1_ejecucion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionEjecucion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoEjecucion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ejecutar el motor para un corte",
        "tags": [
          "ejecucion"
        ]
      }
    },
    "/api/v1/exportar/excel": {
      "get": {
        "operationId": "a_excel_api_v1_exportar_excel_get",
        "parameters": [
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Excel del corte",
        "tags": [
          "exportar"
        ]
      }
    },
    "/api/v1/exportar/pdf": {
      "get": {
        "operationId": "a_pdf_api_v1_exportar_pdf_get",
        "parameters": [
          {
            "in": "query",
            "name": "corte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corte"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "PDF del corte",
        "tags": [
          "exportar"
        ]
      }
    },
...
  },
...
}